+++ examples/headless_render/headless_render_app.cpp
@@ -3,10 +3,11 @@
 #include <cubey/image_io.h>
 #include <cubey/vulkan/buffer.h>
 #include <cubey/vulkan/device.h>
+#include <cubey/vulkan/dynamic_rendering.h>
 #include <cubey/vulkan/image.h>
+#include <cubey/vulkan/image_transitions.h>
 #include <cubey/vulkan/immediate_commands.h>
 #include <cubey/vulkan/instance.h>
-#include <cubey/vulkan/rendering.h>
 #include <cubey/vulkan/vk_check.h>
 
 #include <vulkan/vulkan.h>