+++ src/cubey/vulkan/image_transitions.cpp
@@ -1,4 +1,4 @@
-#include <cubey/vulkan/rendering.h>
+#include <cubey/vulkan/image_transitions.h>
 
 #include <cubey/vulkan/vk_check.h>
 
@@ -144,28 +144,6 @@ void transition_image_layout(VkCommandBuffer command_buffer,
     const VkImageMemoryBarrier barrier = image_memory_barrier(transition);
     vkCmdPipelineBarrier(command_buffer, transition.src_stage_mask, transition.dst_stage_mask, 0, 0,
                          nullptr, 0, nullptr, 1, &barrier);
-}
-
-VkRenderingAttachmentInfo color_rendering_attachment(VkImageView image_view, VkClearValue clear) {
-    auto attachment =
-        vk_struct<VkRenderingAttachmentInfo>(VK_STRUCTURE_TYPE_RENDERING_ATTACHMENT_INFO);
-    attachment.imageView = image_view;
-    attachment.imageLayout = VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL;
-    attachment.loadOp = VK_ATTACHMENT_LOAD_OP_CLEAR;
-    attachment.storeOp = VK_ATTACHMENT_STORE_OP_STORE;
-    attachment.clearValue = clear;
-    return attachment;
-}
-
-VkRenderingAttachmentInfo depth_rendering_attachment(VkImageView image_view, VkClearValue clear) {
-    auto attachment =
-        vk_struct<VkRenderingAttachmentInfo>(VK_STRUCTURE_TYPE_RENDERING_ATTACHMENT_INFO);
-    attachment.imageView = image_view;
-    attachment.imageLayout = VK_IMAGE_LAYOUT_DEPTH_STENCIL_ATTACHMENT_OPTIMAL;
-    attachment.loadOp = VK_ATTACHMENT_LOAD_OP_CLEAR;
-    attachment.storeOp = VK_ATTACHMENT_STORE_OP_DONT_CARE;
-    attachment.clearValue = clear;
-    return attachment;
 }
 
 } // namespace cubey::vulkan